# common-docs/openapi/openapi.v1.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Aini-Inu API",
        "description": "Aini-Inu \ubc31\uc5d4\ub4dc REST API \ubb38\uc11c",
        "contact": {
            "name": "Aini-Inu Backend Team"
        },
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Local"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
    "paths": {
...
        "/api/v1/lost-pets": {
            "get": {
                "tags": [
                    "Lost Pets"
                ],
                "summary": "\uc2e4\uc885 \uc2e0\uace0 \ubaa9\ub85d \uc870\ud68c",
                "description": "\ub0b4 \uc2e4\uc885 \uc2e0\uace0 \ubaa9\ub85d\uc744 \uc0c1\ud0dc \ud544\ud130\uc640 \ud568\uaed8 \uc870\ud68c\ud569\ub2c8\ub2e4.",
                "operationId": "list",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "ACTIVE",
                                "RESOLVED",
                                "CLOSED"
                            ]
                        }
                    },
                    {
                        "name": "pageable",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Pageable"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseSliceResponseLostPetSummaryResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Lost Pets"
                ],
                "summary": "\uc2e4\uc885 \uc2e0\uace0 \uc0dd\uc131",
                "description": "\ubc18\ub824\uacac \uc2e4\uc885 \uc2e0\uace0\ub97c \uc0dd\uc131\ud569\ub2c8\ub2e4.",
                "operationId": "create_2",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LostPetCreateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseLostPetResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/lost-pets/{lostPetId}/match": {
            "get": {
                "tags": [
                    "Lost Pets"
                ],
                "summary": "\ub9e4\uce6d \ud6c4\ubcf4 \uc870\ud68c",
                "description": "\uc138\uc158 \uae30\ubc18 \ub9e4\uce6d \ud6c4\ubcf4\ub97c \uc870\ud68c\ud569\ub2c8\ub2e4. sessionId \ubbf8\uc9c0\uc815 \uc2dc \ucd5c\uc2e0 \uc720\ud6a8 \uc138\uc158 \uae30\uc900\uc785\ub2c8\ub2e4.",
                "operationId": "matchCandidates",
                "parameters": [
                    {
                        "name": "lostPetId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "sessionId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "pageable",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Pageable"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseSliceResponseLostPetMatchCandidateResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Lost Pets"
                ],
                "summary": "\ub9e4\uce6d \ud6c4\ubcf4 \uc2b9\uc778",
                "description": "\uc138\uc158 \ud6c4\ubcf4\ub97c \uc2b9\uc778\ud558\uace0 \ub9e4\uce58/\ucc44\ud305 \uc5f0\uacc4\ub97c \uc2dc\ub3c4\ud569\ub2c8\ub2e4.",
                "operationId": "approveMatch",
                "parameters": [
                    {
                        "name": "lostPetId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LostPetMatchApproveRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "\uc2b9\uc778 \uc131\uacf5",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseLostPetMatchResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "\uc778\uc99d \uc2e4\ud328",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseLostPetMatchResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "\ud0c0\uc778 \uc2e0\uace0 \uc811\uadfc \ubd88\uac00",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseLostPetMatchResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "\ud6c4\ubcf4/\uc0c1\ud0dc \ucda9\ub3cc",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseLostPetMatchResponse"
                                }
                            }
                        }
                    },
                    "410": {
                        "description": "\uc138\uc158 \ub9cc\ub8cc",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseLostPetMatchResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/lost-pets/analyze": {
            "post": {
                "tags": [
                    "Lost Pets"
                ],
                "summary": "\uc2e4\uc885 \uc2e0\uace0 \ubd84\uc11d",
                "description": "\uc2e4\uc885 \uc2e0\uace0 \uc815\ubcf4 \uae30\ubc18 AI \ud6c4\ubcf4 \ud0d0\uc0c9 \uc138\uc158\uc744 \uc0dd\uc131\ud569\ub2c8\ub2e4.",
                "operationId": "analyze",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LostPetAnalyzeRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "\ubd84\uc11d \uc131\uacf5",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseLostPetAnalyzeResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "\uc778\uc99d \uc2e4\ud328",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseLostPetAnalyzeResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "\ud0c0\uc778 \uc2e0\uace0 \uc811\uadfc \ubd88\uac00",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseLostPetAnalyzeResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "AI \ubd84\uc11d \uc2e4\ud328",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseLostPetAnalyzeResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/pets/{petId}": {
            "delete": {
                "tags": [
                    "Pets"
                ],
                "summary": "\ubc18\ub824\uacac \uc0ad\uc81c",
                "description": "\ub4f1\ub85d\ub41c \ubc18\ub824\uacac\uc744 \uc0ad\uc81c\ud569\ub2c8\ub2e4.",
                "operationId": "deletePet",
                "parameters": [
                    {
                        "name": "petId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Pets"
                ],
                "summary": "\ubc18\ub824\uacac \uc218\uc815",
                "description": "\ub4f1\ub85d\ub41c \ubc18\ub824\uacac \uc815\ubcf4\ub97c \ubd80\ubd84 \uc218\uc815\ud569\ub2c8\ub2e4.",
                "operationId": "updatePet",
                "parameters": [
                    {
                        "name": "petId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PetUpdateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponsePetResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/lost-pets/{lostPetId}": {
            "get": {
                "tags": [
                    "Lost Pets"
                ],
                "summary": "\uc2e4\uc885 \uc2e0\uace0 \uc0c1\uc138 \uc870\ud68c",
                "description": "lostPetId \uae30\uc900 \uc2e4\uc885 \uc2e0\uace0 \uc0c1\uc138\ub97c \uc870\ud68c\ud569\ub2c8\ub2e4.",
                "operationId": "detail",
                "parameters": [
                    {
                        "name": "lostPetId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseLostPetDetailResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiResponseVoid": {
                "type": "object",
                "description": "\uacf5\ud1b5 API \uc751\ub2f5 \ub798\ud37c",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "\uc694\uccad \uc131\uacf5 \uc5ec\ubd80",
                        "example": true
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32",
                        "description": "HTTP \uc0c1\ud0dc \ucf54\ub4dc",
                        "example": 200
                    },
                    "data": {
                        "type": "object",
                        "description": "\uc2e4\uc81c \uc751\ub2f5 \ub370\uc774\ud130"
                    },
                    "errorCode": {
                        "type": "string",
                        "description": "\uc5d0\ub7ec \ucf54\ub4dc (\uc131\uacf5 \uc2dc null)",
                        "example": "C101"
                    },
                    "message": {
                        "type": "string",
                        "description": "\uc5d0\ub7ec \uba54\uc2dc\uc9c0 (\uc131\uacf5 \uc2dc null)",
                        "example": "\uc778\uc99d\uc774 \ud544\uc694\ud569\ub2c8\ub2e4"
                    }
                }
            },
...
            "LostPetCreateRequest": {
                "type": "object",
                "properties": {
                    "petName": {
                        "type": "string",
                        "minLength": 1
                    },
                    "breed": {
                        "type": "string"
                    },
                    "photoUrl": {
                        "type": "string",
                        "minLength": 1
                    },
                    "description": {
                        "type": "string"
                    },
                    "lastSeenAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "lastSeenLocation": {
                        "type": "string",
                        "minLength": 1
                    }
                },
                "required": [
                    "lastSeenAt",
                    "lastSeenLocation",
                    "petName",
                    "photoUrl"
                ]
            },
            "ApiResponseLostPetResponse": {
                "type": "object",
                "description": "\uacf5\ud1b5 API \uc751\ub2f5 \ub798\ud37c",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "\uc694\uccad \uc131\uacf5 \uc5ec\ubd80",
                        "example": true
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32",
                        "description": "HTTP \uc0c1\ud0dc \ucf54\ub4dc",
                        "example": 200
                    },
                    "data": {
                        "$ref": "#/components/schemas/LostPetResponse",
                        "description": "\uc2e4\uc81c \uc751\ub2f5 \ub370\uc774\ud130"
                    },
                    "errorCode": {
                        "type": "string",
                        "description": "\uc5d0\ub7ec \ucf54\ub4dc (\uc131\uacf5 \uc2dc null)",
                        "example": "C101"
                    },
                    "message": {
                        "type": "string",
                        "description": "\uc5d0\ub7ec \uba54\uc2dc\uc9c0 (\uc131\uacf5 \uc2dc null)",
                        "example": "\uc778\uc99d\uc774 \ud544\uc694\ud569\ub2c8\ub2e4"
                    }
                }
            },
...
            "LostPetMatchApproveRequest": {
                "type": "object",
                "properties": {
                    "sessionId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "sightingId": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "required": [
                    "sessionId",
                    "sightingId"
                ]
            },
            "ApiResponseLostPetMatchResponse": {
                "type": "object",
                "description": "\uacf5\ud1b5 API \uc751\ub2f5 \ub798\ud37c",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "\uc694\uccad \uc131\uacf5 \uc5ec\ubd80",
                        "example": true
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32",
                        "description": "HTTP \uc0c1\ud0dc \ucf54\ub4dc",
                        "example": 200
                    },
                    "data": {
                        "$ref": "#/components/schemas/LostPetMatchResponse",
                        "description": "\uc2e4\uc81c \uc751\ub2f5 \ub370\uc774\ud130"
                    },
                    "errorCode": {
                        "type": "string",
                        "description": "\uc5d0\ub7ec \ucf54\ub4dc (\uc131\uacf5 \uc2dc null)",
                        "example": "C101"
                    },
                    "message": {
                        "type": "string",
                        "description": "\uc5d0\ub7ec \uba54\uc2dc\uc9c0 (\uc131\uacf5 \uc2dc null)",
                        "example": "\uc778\uc99d\uc774 \ud544\uc694\ud569\ub2c8\ub2e4"
                    }
                }
            },
...
            "LostPetAnalyzeRequest": {
                "type": "object",
                "properties": {
                    "lostPetId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "image": {
                        "type": "string"
                    },
                    "imageUrl": {
                        "type": "string"
                    },
                    "mode": {
                        "type": "string"
                    },
                    "queryText": {
                        "type": "string"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "double"
                    },
                    "longitude": {
                        "type": "number",
                        "format": "double"
                    },
                    "imageProvided": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "lostPetId"
                ]
            },
            "ApiResponseLostPetAnalyzeResponse": {
                "type": "object",
                "description": "\uacf5\ud1b5 API \uc751\ub2f5 \ub798\ud37c",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "\uc694\uccad \uc131\uacf5 \uc5ec\ubd80",
                        "example": true
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32",
                        "description": "HTTP \uc0c1\ud0dc \ucf54\ub4dc",
                        "example": 200
                    },
                    "data": {
                        "$ref": "#/components/schemas/LostPetAnalyzeResponse",
                        "description": "\uc2e4\uc81c \uc751\ub2f5 \ub370\uc774\ud130"
                    },
                    "errorCode": {
                        "type": "string",
                        "description": "\uc5d0\ub7ec \ucf54\ub4dc (\uc131\uacf5 \uc2dc null)",
                        "example": "C101"
                    },
                    "message": {
                        "type": "string",
                        "description": "\uc5d0\ub7ec \uba54\uc2dc\uc9c0 (\uc131\uacf5 \uc2dc null)",
                        "example": "\uc778\uc99d\uc774 \ud544\uc694\ud569\ub2c8\ub2e4"
                    }
                }
            },
...
            "Pageable": {
                "type": "object",
                "properties": {
                    "page": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 0
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 1
                    },
                    "sort": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
            "ApiResponseSliceResponseLostPetSummaryResponse": {
                "type": "object",
                "description": "\uacf5\ud1b5 API \uc751\ub2f5 \ub798\ud37c",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "\uc694\uccad \uc131\uacf5 \uc5ec\ubd80",
                        "example": true
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32",
                        "description": "HTTP \uc0c1\ud0dc \ucf54\ub4dc",
                        "example": 200
                    },
                    "data": {
                        "$ref": "#/components/schemas/SliceResponseLostPetSummaryResponse",
                        "description": "\uc2e4\uc81c \uc751\ub2f5 \ub370\uc774\ud130"
                    },
                    "errorCode": {
                        "type": "string",
                        "description": "\uc5d0\ub7ec \ucf54\ub4dc (\uc131\uacf5 \uc2dc null)",
                        "example": "C101"
                    },
                    "message": {
                        "type": "string",
                        "description": "\uc5d0\ub7ec \uba54\uc2dc\uc9c0 (\uc131\uacf5 \uc2dc null)",
                        "example": "\uc778\uc99d\uc774 \ud544\uc694\ud569\ub2c8\ub2e4"
                    }
                }
            },
...
            "ApiResponseSliceResponseLostPetMatchCandidateResponse": {
                "type": "object",
                "description": "\uacf5\ud1b5 API \uc751\ub2f5 \ub798\ud37c",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "\uc694\uccad \uc131\uacf5 \uc5ec\ubd80",
                        "example": true
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32",
                        "description": "HTTP \uc0c1\ud0dc \ucf54\ub4dc",
                        "example": 200
                    },
                    "data": {
                        "$ref": "#/components/schemas/SliceResponseLostPetMatchCandidateResponse",
                        "description": "\uc2e4\uc81c \uc751\ub2f5 \ub370\uc774\ud130"
                    },
                    "errorCode": {
                        "type": "string",
                        "description": "\uc5d0\ub7ec \ucf54\ub4dc (\uc131\uacf5 \uc2dc null)",
                        "example": "C101"
                    },
                    "message": {
                        "type": "string",
                        "description": "\uc5d0\ub7ec \uba54\uc2dc\uc9c0 (\uc131\uacf5 \uc2dc null)",
                        "example": "\uc778\uc99d\uc774 \ud544\uc694\ud569\ub2c8\ub2e4"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}